# Forward model update and optimizations