--- a/botbowl/core/game.py
+++ b/botbowl/core/game.py
@@ -1615,7 +1615,7 @@ class Game:
                 if from_position.distance(square, manhattan=True) >= 3:
                     include = True
             if include:
-                if self.is_out_of_bounds(square):
+                if square.out_of_bounds:
                     squares_out.append(square)
                 elif self.get_player_at(square) is None:
                     squares_empty.append(square)
@@ -1635,7 +1635,10 @@ class Game:
         :param y:
         :return: A square with the position (x,y)
         """
-        return self.square_shortcut[y][x]
+        try:
+            return self.square_shortcut[y][x]
+        except IndexError:
+            return Square(x, y)
 
     def get_adjacent_squares(self, position: Square, diagonal=True, out=False, occupied=True, distance=1) \
             -> List[Square]:
@@ -1655,7 +1658,7 @@ class Game:
                 if yy == 0 and xx == 0:
                     continue
                 sq = self.get_square(position.x + xx, position.y + yy)
-                if not out and self.is_out_of_bounds(sq):
+                if not out and sq.out_of_bounds:
                     continue
                 if not occupied and self.get_player_at(sq) is not None:
                     continue
@@ -1736,7 +1739,7 @@ class Game:
                 if yy == 0 and xx == 0:
                     continue
                 p = self.get_square(opp_player.position.x + xx, opp_player.position.y + yy)
-                if not self.is_out_of_bounds(p) and player.position != p:
+                if not p.out_of_bounds and player.position != p:
                     player_at = self.get_player_at(p)
                     if player_at is not None:
                         if player_at.team == player.team:
@@ -1773,7 +1776,7 @@ class Game:
                 if yy == 0 and xx == 0:
                     continue
                 p = self.get_square(opp_player.position.x + xx, opp_player.position.y + yy)
-                if not self.is_out_of_bounds(p) and player.position != p:
+                if not p.out_of_bounds and player.position != p:
                     player_at = self.get_player_at(p)
                     if player_at is not None:
                         if player_at.team == player.team and self.can_assist(player_at, foul):
@@ -2087,7 +2090,7 @@ class Game:
         for y in range(len(self.state.pitch.board)):
             for x in range(len(self.state.pitch.board[y])):
                 to_position = self.get_square(x, y)
-                if self.is_out_of_bounds(to_position) or position == to_position:
+                if to_position.out_of_bounds or position == to_position:
                     continue
                 distance = self.get_pass_distance(position, to_position)
                 if distance in distances_allowed: