--- a/botbowl/core/model.py
+++ b/botbowl/core/model.py
@@ -15,7 +15,7 @@ import pickle
 from math import sqrt
 from botbowl.core.util import *
 from botbowl.core.table import *
-from botbowl.core.forward_model import Immutable, Reversible, CallableStep
+from botbowl.core.forward_model import Immutable, Reversible, CallableStep, treat_as_immutable
 
 
 class ReplayStep:
@@ -509,10 +509,10 @@ class ActionChoice(Immutable):
         }
 
 
-class Action(Reversible):
+@treat_as_immutable
+class Action:
 
     def __init__(self, action_type, position=None, player=None):
-        super().__init__()
         self.action_type = action_type
         self.position = position
         self.player = player
@@ -569,10 +569,10 @@ class Die:
         Exception("Method not implemented")
 
 
-class DiceRoll(Reversible):
+@treat_as_immutable
+class DiceRoll:
 
     def __init__(self, dice, modifiers=0, target=None, d68=False, roll_type=RollType.AGILITY_ROLL, target_higher=True, target_lower=False, highest_succeed=True, lowest_fail=True):
-        super().__init__()
         self.dice = dice
         self.sum = 0
         self.d68 = d68
@@ -884,7 +884,7 @@ class Ball(Catchable, Reversible):
                f"is_carried={self.is_carried})"
 
 
-class Bomb(Catchable):
+class Bomb(Catchable, Reversible):
 
     def __init__(self, position, on_ground=True, is_carried=False):
         super().__init__(position, on_ground, is_carried)
@@ -1024,6 +1024,7 @@ class Player(Piece, Reversible):
     def __repr__(self):
         return f"Player(position={self.position if self.position is not None else 'None'}, {self.role.name}, state={self.state})"
 
+
 class Square(Immutable):
 
     def __init__(self, x, y):
@@ -1074,6 +1075,7 @@ class Square(Immutable):
 
     def __repr__(self):
         return f"Square({self.x}, {self.y})"
+
 
 class Race: