--- a/botbowl/core/pathfinding/cython_pathfinding.pyx
+++ b/botbowl/core/pathfinding/cython_pathfinding.pyx
@@ -130,9 +130,6 @@ cdef class Path:
                 self.handoff_roll == other.handoff_roll and \
                 self.foul_roll == other.foul_roll
 
-    def __ne__(self, other):
-        return not (self == other)
-
 # Make the forward model treat Path as an immutable type.
 forward_model.immutable_types.add(Path)