This PR should be considered after PRs #191 and #190 have been merged (they have been merged now). This branch is built on top of those changes. 

This PR contains: 
1. Feature to step the game state forward after being reverted. 
2. New ways to set classes in `models.py` to immutable using decorators instead of inheritence which speeds up execution and readability. 
3. Some general optimizations of forward_model.py, game.py, procedures.py, pathfinding.py