@mrbermell 2021-12-09
I'm going insane, here's some [rubber duck debugging](https://en.wikipedia.org/wiki/Rubber_duck_debugging) mumbling that hopefully makes me solve this. Github test suite _always_ crashes with a Segmentation Fault. It crashes in test_revert_and_forward() in test_forward_model.py while comparing two MoveAction-procedure's paths attributes. Since the paths are from cython_pathfinding.pyx it's harder to debug, it becomes harder and stranger that I only _sometimes_  (<10% of the time) get this fault on my computer. 

More specifically it crashes while calling `assert_game_states` game states for the last time in the loop. That function hasn't changed for a long time. The only change is this tests which now also tests `game.forward(undone_steps)`. But I just don't see it. 

```
============================= test session starts ==============================
platform linux -- Python 3.7.12, pytest-6.2.5, py-1.11.0, pluggy-1.0.0
rootdir: /home/runner/work/botbowl/botbowl
collected 448 items

tests/ai/test_competition.py ..                                          [  0%]
tests/ai/test_env.py .                                                   [  0%]
tests/ai/test_pathfinding.py ........................................... [ 10%]
...........                                                              [ 12%]
Fatal Python error: Segmentation fault

Current thread 0x00007f4f5a0b2740 (most recent call first):
  File "/home/runner/work/botbowl/botbowl/botbowl/core/util.py", line 136 in compare_iterable
  File "/home/runner/work/botbowl/botbowl/botbowl/core/util.py", line 129 in compare_iterable
  File "/home/runner/work/botbowl/botbowl/botbowl/core/util.py", line 157 in compare_object
  File "/home/runner/work/botbowl/botbowl/botbowl/core/procedure.py", line 60 in compare
  File "/home/runner/work/botbowl/botbowl/botbowl/core/util.py", line 122 in compare_iterable
  File "/home/runner/work/botbowl/botbowl/botbowl/core/util.py", line 133 in compare_iterable
  File "/home/runner/work/botbowl/botbowl/botbowl/core/model.py", line 509 in compare
  File "/home/runner/work/botbowl/botbowl/tests/framework/test_forward_model.py", line 152 in assert_game_states
  File "/home/runner/work/botbowl/botbowl/tests/framework/test_forward_model.py", line 31 in test_revert_and_forward
```