@mrbermell 2021-12-11
i had made a mistake when implementing cython_pathfinding.Path.\_\_reduce__() that only showed when a Path without a collected path was pickled. And then, once pickled, it caused a segmentation fault in the c code when it tried to collect the path. 

Anyways, all solved now. And this error is not present in the python implementation because it doesn't use a custom \_\_reduce__()