@njustesen 2021-12-14
Remember to update https://njustesen.github.io/botbowl/forward-model.html