@mrbermell 2022-01-18
@njustesen , I just merged the latest updates from the master branch and updated the tutorial. With that I think it's ready for thorough review.