@mrbermell 2022-01-24
Here's the benchmark as run on [branch before merge](https://github.com/njustesen/botbowl/commit/379d23ff136f54fcb286c2caf812858acfb1b087). 
```
10 games without forward model: 1.27 seconds
10 games with forward model: 1.84 seconds
``` 

Here's the benchmark as run on this branch: 
```
10 games without forward model: 0.76 seconds
10 games with forward model: 1.28 seconds
```

Without forward model enabled it is 1.67x  faster  (1.27/0.76). 

With forward model enabled it's 1.44x  faster  (1.84/1.28).